fix(recap): Fixes the recap summary description
Pelican is pretty naive about how it sets the metadata description
values and just grabs the first text it can find within the HTML. This
can ensnare JavaScript, making the description ugly.

The fix here is twofold: First, we set the length of the summary
explicitly. Second, we move the JavaScript to the end of the document,
where it can't get snagged again in the future.