fix(feeds): Generate feed at build
We were using getServerSideProps, which doesn't seem to
work on netlify, or at least seems to require something
special. On my dev machine, it worked fine, but when
launched, it did nothing. Lame.

So this moves it so that when we generate the blog landing
page, we also generate the atom feed for the blog. It makes
sense, mostly, and comes at small performance penalty. I'm
not sure it'll work in prod, but we'll see.